# MonkeyScheduler.SchedulerService/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MonkeyScheduler API",
    "version": "v1"
  },
  "paths": {
    "/api/dag-workflow/validate": {
      "post": {
        "tags": [
          "DagWorkflow"
        ],
        "parameters": [
          {
            "name": "workflowId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ScheduledTask"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ScheduledTask"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ScheduledTask"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowValidationResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowValidationResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowValidationResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/dag-workflow/detect-cycles": {
      "post": {
        "tags": [
          "DagWorkflow"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ScheduledTask"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ScheduledTask"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ScheduledTask"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CycleDetectionResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CycleDetectionResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CycleDetectionResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/dag-workflow/check-dependencies": {
      "post": {
        "tags": [
          "DagWorkflow"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DependencyCheckRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DependencyCheckRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DependencyCheckRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DependencyCheckResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DependencyCheckResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DependencyCheckResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/dag-workflow/dependency-paths": {
      "post": {
        "tags": [
          "DagWorkflow"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ScheduledTask"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ScheduledTask"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ScheduledTask"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/dag-workflow/start": {
      "post": {
        "tags": [
          "DagWorkflow"
        ],
        "parameters": [
          {
            "name": "workflowId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ScheduledTask"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ScheduledTask"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ScheduledTask"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DagExecutionResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DagExecutionResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DagExecutionResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/dag-workflow/status/{workflowId}": {
      "get": {
        "tags": [
          "DagWorkflow"
        ],
        "parameters": [
          {
            "name": "workflowId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowExecutionStatus"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowExecutionStatus"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowExecutionStatus"
                }
              }
            }
          }
        }
      }
    },
    "/api/dag-workflow/pause/{workflowId}": {
      "post": {
        "tags": [
          "DagWorkflow"
        ],
        "parameters": [
          {
            "name": "workflowId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/dag-workflow/resume/{workflowId}": {
      "post": {
        "tags": [
          "DagWorkflow"
        ],
        "parameters": [
          {
            "name": "workflowId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/dag-workflow/cancel/{workflowId}": {
      "post": {
        "tags": [
          "DagWorkflow"
        ],
        "parameters": [
          {
            "name": "workflowId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CycleDetectionResult": {
        "type": "object",
        "properties": {
          "hasCycle": {
            "type": "boolean"
          },
          "cycles": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            },
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DagExecutionResult": {
        "type": "object",
        "properties": {
          "isSuccess": {
            "type": "boolean"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "startedTaskCount": {
            "type": "integer",
            "format": "int32"
          },
          "workflowId": {
            "type": "string",
            "format": "uuid"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "DagExecutionStatus": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "format": "int32"
      },
      "DependencyCheckRequest": {
        "type": "object",
        "properties": {
          "task": {
            "$ref": "#/components/schemas/ScheduledTask"
          },
          "allTasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduledTask"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DependencyCheckResult": {
        "type": "object",
        "properties": {
          "isValid": {
            "type": "boolean"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "missingDependencies": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "invalidDependencies": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RetryStrategy": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "ScheduledTask": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "cronExpression": {
            "type": "string",
            "nullable": true
          },
          "nextRunTime": {
            "type": "string",
            "format": "date-time"
          },
          "enabled": {
            "type": "boolean"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "taskType": {
            "type": "string",
            "nullable": true
          },
          "taskParameters": {
            "type": "string",
            "nullable": true
          },
          "enableRetry": {
            "type": "boolean"
          },
          "maxRetryCount": {
            "type": "integer",
            "format": "int32"
          },
          "retryIntervalSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "retryStrategy": {
            "$ref": "#/components/schemas/RetryStrategy"
          },
          "currentRetryCount": {
            "type": "integer",
            "format": "int32"
          },
          "nextRetryTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "timeoutSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "dependencies": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "nextTaskIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "dagStatus": {
            "$ref": "#/components/schemas/DagExecutionStatus"
          },
          "completedDependenciesCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalDependenciesCount": {
            "type": "integer",
            "format": "int32"
          },
          "isDagTask": {
            "type": "boolean",
            "readOnly": true
          },
          "dagWorkflowId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "dagPriority": {
            "type": "integer",
            "format": "int32"
          },
          "allowParallelExecution": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "WorkflowExecutionStatus": {
        "type": "object",
        "properties": {
          "workflowId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/WorkflowStatus"
          },
          "totalTasks": {
            "type": "integer",
            "format": "int32"
          },
          "completedTasks": {
            "type": "integer",
            "format": "int32"
          },
          "failedTasks": {
            "type": "integer",
            "format": "int32"
          },
          "runningTasks": {
            "type": "integer",
            "format": "int32"
          },
          "waitingTasks": {
            "type": "integer",
            "format": "int32"
          },
          "skippedTasks": {
            "type": "integer",
            "format": "int32"
          },
          "startTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "endTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "progressPercentage": {
            "type": "number",
            "format": "double",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "WorkflowStatus": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
      "WorkflowValidationResult": {
        "type": "object",
        "properties": {
          "isValid": {
            "type": "boolean"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "entryTasks": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "exitTasks": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "executionLevels": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}